fixed up 404 page, added colors, and created new files for primary button and screengrid
