+++ client/src/Authentication/ResetPasswordPage.tsx
@@ -1,41 +1,32 @@
 import React, { useState } from 'react';
-import { Link, TextField, Button, Grid } from '@mui/material';
-import { useNavigate, Link as RouterLink } from 'react-router-dom';
-
-import { MiniLinkText, FormHeaderText, ScreenGrid } from '../components/grid';
-import FormCol from '../components/form/FormCol';
-import {
-  emailRegex,
-  InputErrorMessage,
-  nameRegex,
-  passwordRegex,
-} from '../util/inputvalidation';
-import { register } from './api';
+import { TextField, Link, Button, Grid } from '@mui/material';
+import { useNavigate, useParams, Link as RouterLink } from 'react-router-dom';
+import { MiniLinkText } from '../components/grid';
+import { resetPassword } from './api';
+import FormGrid from '../components/form/FormGrid';
 import AlertDialog from '../components/AlertDialog';
-
-function RegisterPage() {
+import { InputErrorMessage, passwordRegex } from '../util/inputvalidation';
+import PrimaryButton from '../components/buttons/PrimaryButton';
+import Gridd from '../components/ScreenGrid';
+/**
+ * A page that allows users to reset their password by inputting a new password
+ * into a form.
+ */
+function ResetPasswordPage() {
+  const { token } = useParams();
   const navigate = useNavigate();
 
   // Default values for state
   const defaultValues = {
-    firstName: '',
-    lastName: '',
-    email: '',
     password: '',
     confirmPassword: '',
   };
   const defaultShowErrors = {
-    firstName: false,
-    lastName: false,
-    email: false,
     password: false,
     confirmPassword: false,
     alert: false,
   };
   const defaultErrorMessages = {
-    firstName: '',
-    lastName: '',
-    email: '',
     password: '',
     confirmPassword: '',
     alert: '',
@@ -46,8 +37,6 @@ function RegisterPage() {
   const [values, setValueState] = useState(defaultValues);
   const [showError, setShowErrorState] = useState(defaultShowErrors);
   const [errorMessage, setErrorMessageState] = useState(defaultErrorMessages);
-  const [alertTitle, setAlertTitle] = useState('Error');
-  const [isRegistered, setRegistered] = useState(false);
 
   // Helper functions for changing only one field in a state object
   const setValue = (field: string, value: string) => {
@@ -69,13 +58,6 @@ function RegisterPage() {
     }));
   };
 
-  const handleAlertClose = () => {
-    if (isRegistered) {
-      navigate('/login');
-    }
-    setShowError('alert', false);
-  };
-
   const clearErrorMessages = () => {
     setShowErrorState(defaultShowErrors);
     setErrorMessageState(defaultErrorMessages);
@@ -95,98 +77,47 @@ function RegisterPage() {
       }
     }
 
-    if (!values.firstName.match(nameRegex)) {
-      setErrorMessage('firstName', InputErrorMessage.INVALID_NAME);
-      setShowError('firstName', true);
-      isValid = false;
-    }
-    if (!values.lastName.match(nameRegex)) {
-      setErrorMessage('lastName', InputErrorMessage.INVALID_NAME);
-      setShowError('lastName', true);
-      isValid = false;
-    }
-    if (!values.email.match(emailRegex)) {
-      setErrorMessage('email', InputErrorMessage.INVALID_EMAIL);
-      setShowError('email', true);
-      isValid = false;
-    }
     if (!values.password.match(passwordRegex)) {
       setErrorMessage('password', InputErrorMessage.INVALID_PASSWORD);
       setShowError('password', true);
       isValid = false;
     }
-    if (!(values.confirmPassword === values.password)) {
+    if (!(values.password === values.confirmPassword)) {
       setErrorMessage('confirmPassword', InputErrorMessage.PASSWORD_MISMATCH);
       setShowError('confirmPassword', true);
       isValid = false;
     }
-
     return isValid;
+  };
+
+  const alertTitle = 'Error';
+  const handleAlertClose = () => {
+    setShowError('alert', false);
   };
 
   async function handleSubmit() {
     if (validateInputs()) {
-      register(values.firstName, values.lastName, values.email, values.password)
+      resetPassword(values.password, token || 'missing token')
         .then(() => {
-          setShowError('alert', true);
-          setAlertTitle('');
-          setRegistered(true);
-          setErrorMessage('alert', 'Check email to verify account');
+          navigate('/');
         })
         .catch((e) => {
-          setShowError('alert', true);
           setErrorMessage('alert', e.message);
+          setShowError('alert', true);
         });
     }
   }
 
   return (
-    <ScreenGrid>
-      <FormCol>
-        <Grid item>
-          <TextField
-            error={showError.firstName}
-            helperText={errorMessage.firstName}
-            size="small"
-            type="text"
-            required
-            label="First Name"
-            value={values.firstName}
-            onChange={(e) => setValue('firstName', e.target.value)}
-          />
-        </Grid>
-        <Grid item>
-          <TextField
-            error={showError.lastName}
-            helperText={errorMessage.lastName}
-            size="small"
-            type="text"
-            required
-            label="Last Name"
-            value={values.lastName}
-            onChange={(e) => setValue('lastName', e.target.value)}
-          />
-        </Grid>
-        <Grid item>
-          <TextField
-            error={showError.email}
-            helperText={errorMessage.email}
-            size="small"
-            type="text"
-            required
-            label="Email"
-            value={values.email}
-            onChange={(e) => setValue('email', e.target.value)}
-          />
-        </Grid>
+    <Gridd>
+      <FormGrid>
         <Grid item>
           <TextField
             error={showError.password}
             helperText={errorMessage.password}
-            size="small"
             type="password"
             required
-            label="Password"
+            label="New Password"
             value={values.password}
             onChange={(e) => setValue('password', e.target.value)}
           />
@@ -195,23 +126,22 @@ function RegisterPage() {
           <TextField
             error={showError.confirmPassword}
             helperText={errorMessage.confirmPassword}
-            size="small"
             type="password"
             required
-            label=" Confirm Password"
+            label="Confirm Password"
             value={values.confirmPassword}
             onChange={(e) => setValue('confirmPassword', e.target.value)}
           />
         </Grid>
         <Grid item>
-          <Button
+          <PrimaryButton
             type="submit"
             variant="contained"
             color="primary"
             onClick={() => handleSubmit()}
           >
-            Register
-          </Button>
+            Reset Password
+          </PrimaryButton>
         </Grid>
         <Grid item>
           <MiniLinkText>
@@ -221,18 +151,16 @@ function RegisterPage() {
             </Link>
           </MiniLinkText>
         </Grid>
-      </FormCol>
+      </FormGrid>
       {/* The alert that pops up */}
-      <Grid item>
-        <AlertDialog
-          showAlert={showError.alert}
-          title={alertTitle}
-          message={errorMessage.alert}
-          onClose={handleAlertClose}
-        />
-      </Grid>
-    </ScreenGrid>
+      <AlertDialog
+        showAlert={showError.alert}
+        title={alertTitle}
+        message={errorMessage.alert}
+        onClose={handleAlertClose}
+      />
+    </Gridd>
   );
 }
 
-export default RegisterPage;
+export default ResetPasswordPage;
