+++ client/src/Authentication/VerifyAccountPage.tsx
@@ -1,8 +1,9 @@
 import React, { useState, useEffect } from 'react';
 import { Link, CircularProgress, Grid, Typography } from '@mui/material';
 import { useParams, Link as RouterLink } from 'react-router-dom';
-import { MiniLinkText, ScreenGrid } from '../components/grid';
+import { MiniLinkText } from '../components/grid';
 import { verifyAccount } from './api';
+import Gridd from '../components/ScreenGrid';
 
 /**
  * A page users visit to verify their account. Page should be accessed via
@@ -28,11 +29,11 @@ function VerifyAccountPage() {
   }, [token]); // Only runs when there is a change in token
 
   return loading ? (
-    <ScreenGrid>
+    <Gridd>
       <CircularProgress />
-    </ScreenGrid>
+    </Gridd>
   ) : (
-    <ScreenGrid>
+    <Gridd>
       <Grid item>
         <Typography variant="h5">{message}</Typography>
       </Grid>
@@ -44,7 +45,7 @@ function VerifyAccountPage() {
           </Link>
         </MiniLinkText>
       </Grid>
-    </ScreenGrid>
+    </Gridd>
   );
 }
