+++ client/src/assets/colors.ts
@@ -1,6 +1,20 @@
 const COLORS = {
-  primary: '#065798',
-  secondary: '#83E4C7',
-  // your colors
+  // From Official hack website
+  // https://github.com/hack4impact/chapter-website-template/blob/main/public/style.css
+  primaryBlue: '#0069ca',
+  secondarySeafoam: '#80d2c8',
+  gray: '#657788',
+  primaryDark: '#172a3a', // Dark Blue
+  primaryDark2: '#2e5987', // Dark Blue level 2
+  background: '#d7fdff', // Light Light Blue
+  accentOrange: '#f2594b',
+  black: '#171717', // Black
+  white: '#ffffff', // White
+  neutral: '#8c8c8c', // Gray
+  lightGray: '#ebeff2', // Gray-subsection
+  actionBlue: '#155da1', // Action button
+  // TODO: get the others colors not at the top of file also in
+
+  // Other colors:
 };
 export default COLORS;
