--- client/src/assets/fonts.ts
@@ -1,9 +1,0 @@
-const FONTS = {
-  // your fonts
-  roboto: {
-    headerFont: 'Roboto',
-    headerWeight: 'bold',
-  },
-};
-
-export default FONTS;
